requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.0.8
django-crispy-forms==1.9.1
django-filter==2.3.0
gunicorn==20.0.4
heroku==0.1.4
idna==2.10
Pillow==7.2.0
python-dateutil==1.5
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.1.0
